requirements.txt
pyyaml>=6.0
anthropic>=0.49.0
requests>=2.31.0
feedparser>=6.0.0
